fix(chat): tighten status and compaction followups
